# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comic Canvas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cast": {
      "get": {
        "tags": [
          "cast"
        ],
        "summary": "List Cast",
        "operationId": "list_cast_cast_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Cast Cast Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "Create Person",
        "operationId": "create_person_cast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Person Cast Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cast/{person_id}": {
      "patch": {
        "tags": [
          "cast"
        ],
        "summary": "Update Person",
        "operationId": "update_person_cast__person_id__patch",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonPatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Person Cast  Person Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cast"
        ],
        "summary": "Revoke Person",
        "operationId": "revoke_person_cast__person_id__delete",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Revoke Person Cast  Person Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cast/{person_id}/photos/presign": {
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "Presign Photo",
        "operationId": "presign_photo_cast__person_id__photos_presign_post",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoPresignIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Presign Photo Cast  Person Id  Photos Presign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cast/{person_id}/process": {
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "Process Person",
        "operationId": "process_person_cast__person_id__process_post",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Process Person Cast  Person Id  Process Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cast/{person_id}/train": {
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "Train Person",
        "description": "The worker's `train_character_job` builds the look card (task\n5.4's real vision call, needs `app.llm` \u2014 worker-only, see this\nrouter's module docstring) before running `design_master` itself \u2014\nnot done here in `/process` even though it would be more responsive,\nbecause `ml.identity.lookcard`'s `services/worker` sys.path bootstrap\nbreaks once THIS service's own `app` package is already loaded\nfirst (a real conflict, found live running this router for the first\ntime \u2014 `ModuleNotFoundError: No module named 'app.llm'`).",
        "operationId": "train_person_cast__person_id__train_post",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Train Person Cast  Person Id  Train Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cast/{person_id}/master": {
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "Pick Master",
        "description": "The person's own choice of master (task 5.18's \"which one feels\nlike you\" screen, user request 2026-09-19) \u2014 promotes one of the\ncandidates `train_character_job` saved to `master.png`, records the\nsame provenance fields the old auto-approve wrote, and discards the\nrest. Free: no generation happens here.",
        "operationId": "pick_master_cast__person_id__master_post",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterPickIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Pick Master Cast  Person Id  Master Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cast/invite": {
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "Create Invite",
        "description": "A signed, stateless token (no `invites` table exists in CLAUDE.md\n\u00a73) \u2014 `expires`+`user_id` are embedded and HMAC-signed with the same\nlocal secret `LocalFileStorage`'s own upload tokens use, verified the\nsame way (`hmac.compare_digest`).",
        "operationId": "create_invite_cast_invite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Invite Cast Invite Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InviteIn": {
        "properties": {
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 604800
          }
        },
        "type": "object",
        "title": "InviteIn"
      },
      "MasterPickIn": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          }
        },
        "type": "object",
        "required": [
          "candidate_id"
        ],
        "title": "MasterPickIn"
      },
...
      "PersonIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender_term": {
            "type": "string",
            "title": "Gender Term",
            "default": ""
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PersonIn"
      },
      "PersonPatchIn": {
        "properties": {
          "gender_term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender Term"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          }
        },
        "type": "object",
        "title": "PersonPatchIn",
        "description": "A person is auto-created (just a name) the moment the setup\nwizard's first screen loads, before there's been any chance to ask\nfor a gender \u2014 this lets that same screen fill it in right after,\nrather than requiring person-creation itself to block on it. Both\nfields are optional so the wizard can save gender and age\nindependently (`COALESCE` below keeps whichever isn't sent)."
      },
      "PhotoPresignIn": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "image/jpeg"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "PhotoPresignIn"
      },
...
      "TrainIn": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          }
        },
        "type": "object",
        "title": "TrainIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}